tests/wpt/web-platform-tests/tools/wptrunner/test/testdata/testharness/testharness_0.html

<!doctype html>
<title>Test should be disabled</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(false);
}, "Test that should fail");
</script>
